consolidate shared and editor
