+++ src/components/Editor/Footer.js
@@ -2,17 +2,17 @@ const React = require("react");
 const { DOM: dom, PropTypes } = React;
 const { connect } = require("react-redux");
 const { bindActionCreators } = require("redux");
-const actions = require("../actions");
+const actions = require("../../actions");
 const { getSelectedSource, getSourceText,
-        getPrettySource, getPaneCollapse } = require("../selectors");
-const Svg = require("./shared/Svg");
+        getPrettySource, getPaneCollapse } = require("../../selectors");
+const Svg = require("../shared/Svg");
 const ImPropTypes = require("react-immutable-proptypes");
 const classnames = require("classnames");
 const { isEnabled } = require("devtools-config");
-const { isPretty } = require("../utils/source");
-const { shouldShowFooter, shouldShowPrettyPrint } = require("../utils/editor");
+const { isPretty } = require("../../utils/source");
+const { shouldShowFooter, shouldShowPrettyPrint } = require("../../utils/editor");
 const PaneToggleButton = React.createFactory(
-  require("./shared/Button/PaneToggle")
+  require("../shared/Button/PaneToggle")
 );
 
 require("./Footer.css");
