+++ src/components/Editor/index.js
@@ -11,8 +11,8 @@ const SourceEditor = require("../../utils/source-editor");
 const { find, findNext, findPrev, removeOverlay } = require("../../utils/source-search");
 const { getMode } = require("../../utils/source");
 const Footer = createFactory(require("./Footer"));
-const SearchBar = createFactory(require("./Editor/SearchBar"));
-const { renderConditionalPanel } = require("./Editor/ConditionalPanel");
+const SearchBar = createFactory(require("./SearchBar"));
+const { renderConditionalPanel } = require("./ConditionalPanel");
 const { debugGlobal } = require("devtools-launchpad");
 const {
   getSourceText, getBreakpointsForSource,
@@ -22,13 +22,13 @@ const {
 } = require("../../selectors");
 const { makeLocationId } = require("../../reducers/breakpoints");
 const actions = require("../../actions");
-const Breakpoint = React.createFactory(require("./Editor/Breakpoint"));
-const HitMarker = React.createFactory(require("./Editor/HitMarker"));
+const Breakpoint = React.createFactory(require("./Breakpoint"));
+const HitMarker = React.createFactory(require("./HitMarker"));
 
 const { getDocument, setDocument } = require("../../utils/source-documents");
 const { shouldShowFooter, clearLineClass, onKeyDown } = require("../../utils/editor");
 const { isFirefox } = require("devtools-config");
-const { showMenu } = require("../../utils/menu");
+const { showMenu } = require("../shared/menu");
 const { isEnabled } = require("devtools-config");
 const { isOriginalId, hasMappedSource } = require("../../utils/source-map");
